requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.16              # Vision Transformer (ViT) 지원

# ── Data Science ──────────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0        # GroupKFold, metrics
scipy>=1.11.0              # Calibration 등

# ── Image Processing ──────────────────────────────────────────────────────────
Pillow>=10.0.0
opencv-python-headless>=4.8.0  # cv2 CLAHE (headless: GUI 불필요)
pydicom>=2.4.0             # DICOM 메타데이터 파싱
albumentations>=1.3.1      # 고급 Augmentation

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
plotly>=5.18.0
seaborn>=0.12.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
httpx>=0.27.0              # pytest용 비동기 HTTP 클라이언트

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.33.0
requests>=2.31.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.29.0
